Remove async integration tests
