--- .env.integration_async
@@ -1,20 +1,0 @@
-KERNEL_CLASS='App\Kernel'
-APP_SECRET='$ecretf0rt3st'
-SYMFONY_DEPRECATIONS_HELPER=999999
-
-COMPILER_HOST=localhost
-COMPILER_PORT=9500
-COMPILER_SOURCE_DIRECTORY=/app/source
-COMPILER_TARGET_DIRECTORY=/app/tests
-
-DELEGATOR_HOST=localhost
-DELEGATOR_PORT=9501
-
-HTTP_LOG_PATH=tests/log/http_log.txt
-CALLBACK_RETRY_LIMIT=3
-JOB_TIMEOUT_CHECK_PERIOD_MS=1000
-
-CALLBACK_BASE_URL=http://localhost:8080
-BASIL_LOCAL_FIXTURES_PATH=tests/Fixtures/Basil
-BASIL_LOCAL_UPLOAD_PATH=tests/Fixtures/Uploaded
-BASIL_LOCAL_SOURCE_PATH=tests/Fixtures/CompilerSource
