+++ .github/workflows/tests-integration.yml
@@ -34,72 +34,36 @@ jobs:
         run: composer integration-setup
 
       - name: Run tests
-        run: composer test-integration-sync
-
-  integration-async:
-    name: Async integration tests - PHP ${{ matrix.php-versions }} ${{ matrix.operating-system }}
-    strategy:
-      matrix:
-        operating-system: [ubuntu-latest]
-        php-versions: ['8']
-    runs-on: ${{ matrix.operating-system }}
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-
-      - name: Setup php
-        uses: shivammathur/setup-php@v2
-        with:
-          php-version: ${{ matrix.php-versions }}
-          extensions: amqp, pgsql
-
-      - name: Download dependencies
-        uses: php-actions/composer@v2
-        with:
-          args: --ignore-platform-reqs --quiet
-
-      - name: Setup integration test environment
-        run: composer integration-setup
-
-      - name: Run tests
-        run: |
-          php bin/console --env=integration_async messenger:consume async > /dev/null &
-          composer test-integration-async
-
-  integration-app:
-    name: App integration tests - PHP ${{ matrix.php-versions }} ${{ matrix.operating-system }}
-    strategy:
-      matrix:
-        operating-system: [ubuntu-latest]
-        php-versions: ['8']
-    runs-on: ${{ matrix.operating-system }}
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-
-      - name: Setup php
-        uses: shivammathur/setup-php@v2
-        with:
-          php-version: ${{ matrix.php-versions }}
-          extensions: amqp, pgsql
-
-      - name: Download dependencies
-        uses: php-actions/composer@v2
-        with:
-          args: --ignore-platform-reqs --quiet
-
-      - name: Setup integration test environment
-        run: |
-          composer integration-setup
-          composer integration-app-start
-          composer integration-chown-app-source
-
-      - name: Run tests
         run: composer test-integration
 
-      - name: Teardown integration test environment
+  integration-app:
+    name: Image integration tests - PHP ${{ matrix.php-versions }} ${{ matrix.operating-system }}
+    strategy:
+      matrix:
+        operating-system: [ubuntu-latest]
+        php-versions: ['8']
+    runs-on: ${{ matrix.operating-system }}
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+
+      - name: Setup php
+        uses: shivammathur/setup-php@v2
+        with:
+          php-version: ${{ matrix.php-versions }}
+          extensions: amqp, pgsql
+
+      - name: Download dependencies
+        uses: php-actions/composer@v2
+        with:
+          args: --ignore-platform-reqs --quiet
+
+      - name: Setup image test environment
         run: |
-          sudo composer integration-teardown
-          sudo rm -Rf $PWD/var/basil/local/tests/*
+          composer integration-setup
+          composer image-start
+          composer image-chown-app-source
+
+      - name: Run tests
+        run: composer test-image
