+++ composer.json
@@ -121,12 +121,12 @@
             "LOCAL_SOURCE_PATH=\"$PWD/tests/Fixtures/CompilerSource\" docker-compose --env-file ./.docker-compose.env -f ./docker-compose.yml -f ./tests/docker-compose-overrides.yml up -d --build",
             "docker-compose --env-file ./.docker-compose.env stop app-web"
         ],
-        "integration-app-start": [
+        "image-start": [
             "docker-compose --env-file ./.docker-compose.env stop app-web",
             "@db-recreate",
             "docker-compose --env-file ./.docker-compose.env start app-web"
         ],
-        "integration-chown-app-source": [
+        "image-chown-app-source": [
             "docker-compose --env-file ./.docker-compose.env exec -T app-web chown -R www-data:www-data /app/source"
         ],
         "integration-status": "docker-compose --env-file ./.docker-compose.env -f ./docker-compose.yml -f ./tests/docker-compose-overrides.yml ps",
@@ -138,11 +138,10 @@
             "sleep 2",
             "@db-setup"
         ],
-        "test-integration": [
-            "./vendor/bin/phpunit tests/Integration/AppTest.php"
+        "test-image": [
+            "./vendor/bin/phpunit tests/Image/AppTest.php"
         ],
-        "test-integration-async": "./vendor/bin/phpunit -c phpunit.integration_async.xml",
-        "test-integration-sync": "./vendor/bin/phpunit -c phpunit.integration_sync.xml",
+        "test-integration-sync": "./vendor/bin/phpunit -c phpunit.integration.xml",
         "ci": [
             "@cs",
             "@static-analysis",
