--- config/packages/integration_async/framework.yaml
@@ -1,4 +1,0 @@
-framework:
-    test: true
-    session:
-        storage_id: session.storage.mock_file
