--- config/services_integration_async.yaml
@@ -1,16 +1,0 @@
-imports:
-    - { resource: services/test_common.yaml }
-    - { resource: services/integration_common.yaml }
-
-services:
-    _defaults:
-        autowire: true
-        autoconfigure: true
-
-    App\Services\BackoffStrategyFactory:
-        class: App\Tests\Services\Integration\BackoffStrategyFactory
-        arguments:
-            $maximumNumberOfDelays: '%callback_retry_limit%'
-
-    App\Tests\Services\EntityRetriever:
-        public: true
