--- phpunit.integration_sync.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- https://phpunit.readthedocs.io/en/latest/configuration.html -->
-<phpunit
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/9.3/phpunit.xsd"
-        backupGlobals="false"
-        bootstrap="tests/bootstrap.php"
-        colors="true"
-        convertErrorsToExceptions="true"
-        convertNoticesToExceptions="true"
-        convertWarningsToExceptions="true"
-        stopOnError="true"
-        stopOnFailure="true"
->
-  <coverage processUncoveredFiles="true">
-    <include>
-      <directory suffix=".php">src</directory>
-    </include>
-  </coverage>
-  <php>
-    <ini name="error_reporting" value="-1"/>
-    <server name="APP_ENV" value="integration_sync" force="true"/>
-    <server name="SHELL_VERBOSITY" value="-1"/>
-  </php>
-  <testsuites>
-    <testsuite name="Project Test Suite">
-      <directory>tests/Integration/Synchronous</directory>
-    </testsuite>
-  </testsuites>
-  <extensions>
-    <extension class="DAMA\DoctrineTestBundle\PHPUnit\PHPUnitExtension" />
-  </extensions>
-</phpunit>
