+++ tests/Integration/AbstractBaseIntegrationTest.php
@@ -6,10 +6,13 @@ namespace App\Tests\Integration;
 
 use App\Tests\AbstractBaseFunctionalTest;
 use App\Tests\Services\EntityRemover;
+use App\Tests\Services\FileStoreHandler;
 
 abstract class AbstractBaseIntegrationTest extends AbstractBaseFunctionalTest
 {
     protected EntityRemover $entityRemover;
+    protected FileStoreHandler $localSourceStoreHandler;
+    protected FileStoreHandler $uploadStoreHandler;
 
     protected function setUp(): void
     {
@@ -18,5 +21,29 @@ abstract class AbstractBaseIntegrationTest extends AbstractBaseFunctionalTest
         $entityRemover = self::$container->get(EntityRemover::class);
         \assert($entityRemover instanceof EntityRemover);
         $this->entityRemover = $entityRemover;
+
+        $localSourceStoreHandler = self::$container->get('app.tests.services.file_store_handler.local_source');
+        \assert($localSourceStoreHandler instanceof FileStoreHandler);
+        $this->localSourceStoreHandler = $localSourceStoreHandler;
+
+        $uploadStoreHandler = self::$container->get('app.tests.services.file_store_handler.uploaded');
+        \assert($uploadStoreHandler instanceof FileStoreHandler);
+        $this->uploadStoreHandler = $uploadStoreHandler;
+
+        $this->clear();
+    }
+
+    protected function tearDown(): void
+    {
+        $this->clear();
+
+        parent::tearDown();
+    }
+
+    private function clear(): void
+    {
+        $this->entityRemover->removeAll();
+        $this->localSourceStoreHandler->clear();
+        $this->uploadStoreHandler->clear();
     }
 }
